Create utility class for compareXml/compareTxt for refactoring purposes
